ledger-api/pyproject.toml
[project]
name = "ledger-api"
version = "1.0.0"
description = "Ledger API service for quantum-resonance - single source of truth for transactions, allocations, and treasury"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.24.0",
    "sqlalchemy==2.0.36",
    "alembic==1.14.0",
    "pydantic==2.10.3",
    "pydantic-settings==2.6.1",
    "asyncpg==0.30.0",
    "pytest==8.3.4",
    "pytest-asyncio==0.24.0",
    "httpx==0.28.1",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-dotenv==1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest-cov==6.0.0",
    "black==24.10.0",
    "ruff==0.8.4",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
